use-cases/post-sales-chatbot/api/requirements.txt
fastapi==0.124.0
uvicorn[standard]==0.38.0
httpx==0.28.1
python-dotenv==1.2.1
pandas==2.3.3
sap-ai-sdk-gen==5.9.0
pydantic==2.12.5
langgraph==1.0.1
langchain==0.3.27
langchain-core==0.3.80
langchain-openai==0.3.35
